+++ jobs/tests/test_utils.py
@@ -739,7 +739,9 @@ WRONG_NIGHTLY_METADATA_FILENAMES = [
     ('firefox', 'pub/firefox/nightly/latest-mozilla-central/firefox-58.0a1.en-US.'
                 'win64.test_packages.json'),
     ('firefox', 'pub/firefox/nightly/2017/11/2017-11-29-11-10-30-mozilla-central/'
-                'firefox-59.0a1.en-US.win32.test_packages.json')
+                'firefox-59.0a1.en-US.win32.test_packages.json'),
+    ('firefox', 'pub/firefox/nightly/2017/12/2017-12-19-10-02-03-mozilla-central/'
+                'firefox-59.0a1.en-US.win32.mozinfo.json'),
 ]