+++ jobs/CHANGELOG.rst
@@ -7,6 +7,7 @@ This document describes changes between each past release.
 ------------------
 
 - Fix exclusion of thunderbird nightly releases (fixes #312)
+- Prevent mozinfo JSON files to be mistaken as Nightly metadata (fixes #315)
 
 1.1.1 (2017-11-30)
 ------------------